requirements.txt
Flask
numpy
joblib
absl-py
astunparse
regex
tensorflow-cpu == 2.5.0
pywebio == 1.3.0